Remove unnecessary cleverness from Speech model
It is not currently possible to set a second organisation on a speech
because of some cleverness in the model that automatically overrides
organisations to that of the person who delivered the speech. This is
unnecessary cleverness - we should allow editors to set the
organisation(s) explicitly as with all other editions.

This required a small change to the behaviour of the speech importer
to set the lead organisation based on the import row (previously it
was relying on the lead organisation being inferred from the role
appointment set during speed tagging)

Verified with Neil Williams.