Fixed email logic that infers create or update
When a first major edition is published we
enqueue a job to alert subscibers about it.
An editor updated a freshly published edition
while this job was in queue and marked it as a
minor change. Due to an bug in the change history
generation logic, fixed in (http://git.io/Cds89w)
it created a change history entry for this minor
change as well.

So when this job was picked up for processing, it
added a line to indicate that an update happened
because it showed 2 change history entries.

Updated the logic to use a better way to determine
if the email is for a create or update, without
infering it from change history.