Use new slave DB config for export.rake
Use a new section in the database config to select the hostname of the MySQL
slave when running `export:*` rake tasks. This allows the hostname to be
controlled by alphagov-deployment and set differently on Platform1, where
Whitehall has it's own MySQL master/slave.

PRs gds/alphagov-deployment#471 and gds/alphagov-deployment#472 must be
merged and promoted to `release` first.

Story in Infra's Pivotal has more information:
https://www.pivotaltracker.com/story/show/66419616