Improve argument checking in topic retagging tasks
If an argument is missed, report what the missed argument was.
When reporting problems, send the output to stderr so it doesn't interfere with other output.
When there's an error, report a non-zero exit status.
If the destination file already present, report an error so we don't accidentally overwrite files.