Stop HtmlAttachment slug generation complaining
friendly_id doesn't like generating slugs for HtmlAttachment because it will use
the scope on the base class Attachment, and that doesn't have the necessary
friendly_id helpers mixed in (because it doesn't do slugs).

By specifying the scope here explicitly, we force friendly_id to go through the
correct class. It also means that we HtmlAttachment slugs are now scoped to
their owning edition, meaning that two HtmlAttachments can share the same slug
assuming they are attached to a different edition.