Patch bug in Rails routing
Whitehall.atom_feed_maker was having trouble generating urls for non-localised
resources, e.g. topics. This turned out to be due to a bug in Rails routing code
that meant the default format set in atom_feed_maker was being lost. This
patches Rails to fix the issue. There is also a pull request on Rails to fix it
there too:

https://github.com/rails/rails/pull/18627