Maintain ordering of inapplicable nations
The forms for editions display the potentially inapplicable nations in a set
order (Scotland, Wales, Northern Ireland). This was previously maintained by
building NationInapplicabilities in the controller and sorting the association
based on the nations. This no longer works in Rails 4 (the order of items in an
association cannot be manipulated in this way). Instead of doing this work in
the controller, we can build the form fields starting from the ordered nations.