Fix issue with invalid AttachmentData
The upgrade to Rails 4 broke mass-assignment of FileAttachments whilst the
associated Attachment file was awaiting virus scanning. The fix is to ensure
that an AttachmentData record will still be valid whilst awaiting virus
scanning.

This change is required because ActiveRecord now loads the associated
AttachmentData *before* it checks if the nested attributes should be assigned or
rejected because of the reject_if proc. This made the FileAttachment invalid for
newly-uploaded FilAttachments that are still awaiting virus scanning because of
the presence check on the AttachmentData `file` attribute.