Use latest shared_mustache
This newer version of shared_mustache gets around an issue where the latest
sprockets-rails was using the blank vendor version of template.js in
shared_mustache instead of the local copy in whitehall when loading non-compiled
assets in non-production environments. This was causing a javascript-enabled
cucumber feature to fail. Because template.js is no longer present in
shared_mustache, we now need a copy in the repository.