Rename "locale" to "primary_locale" on editions
The fact that this column was called "locale" was conflicting with the latest
Rails-4 compatible version of globalize and was causing issues when trying
to save invalid translations.

"primary_locale" is a more suitable name for it. It is "en" by default in the
database and can be overridden for WorldLocationNewsArticles when the article is
for a non-English-speaking audience and therefor is not to have an English
translation.