Replace "locale" column with "primary_locale"
The fact that we are using a column called "locale" on the editions table to
signify the locale for non-english editions causes an issue with later
versions of the globalize gem. This is the first phase of migrating this column
to a new name of "primary_locale". To prevent issues with the database schema
and the running rails processes column information being out-of-sync, we add a
new column and deprecate the old rather than simply renaming it. Once this code
has been deployed, the pending migration can be run to safely drop the old
column and the `deprecated_columns` directive can be removed.