Update non-english locale fields  JavaScript
The DOM selector changed when the locale column was renamed. Whilst I was here I
also tidied up the tests for the JS and changed the behaviour slightly so that
instead of hiding the locale fields on-load, the "js-hidden" class is applied to
the fieldset so that they are hidden right away (on JS-enabled browsers at
least). This prevents them briefly flashing up whilst the JavaScript is being
loaded and evaluated on the page.