Add note regarding globalize fork
@heathd has got a patch in to the main globalize fork that deals with the issue
my fork works around. Unfortunately, master branch globalize has been updated
for Rails 4.2 compatibility and is therefor not compatible with this version of
whitehall. Either the patch will need to be backported, or we'll have to wait
until whitehall is updated for Rails 4.2 before moving to the main fork of
globalize.