Override association setter to handle Rails bug
There is an obscure bug in Rails that will cause db integrity exceptions to be
raised under a very specific set of circumstances (involving a has_many through
association that has accepts_nested_attributes_for on it as well as DB
constraints on the foreign keys). By overriding the setter on the association,
we are able to work around this bug.

There is an open pull request for this[1]. Once the fix is backported to Rails
4.1 (or possibly 4.2), this code should be removed.

[1] https://github.com/rails/rails/pull/18198