Removed block of content referencing policies
During the period of government forming policy is in a state of flux. We are also releasing a new format to the content and this content is not entirely appropriate.

A section discussing policy will be reinstated at the same time that we correct the page after the new Prime Minister and the nature of the government is known.