Fix #document_url for statistics
Because `#document_url` checked the class of the edition to see if it's
a `Publication`, it fails when the edition has been wrapped in a
presenter. This is the case in the index action of the
`StatisticsController`, which wrap all the editions in a
`PublicationesquePresenter`.

To address this, we call `edition.statistics?` without checking the type
first and define a sensible default (`false`) for the method in the base
class.