Set public timestamp when publishing policy papers.
This sets the public timestamp to 24 hours ago
in order to bring the policy paper to the top of
the policy finder after the creation of the new
government.

Although the date doesn't need to be 24 hours ago
in order for it to be associated with the correct
government (because we're setting the public_timestamp),
this is done as the date is visible to users and should
be consistent with reality.